+++ sphinx/application/common/wrappers/wrapper-common/src/main/java/chat/sphinx/wrapper_common/dashboard/ContactId.kt
@@ -9,7 +9,9 @@ inline fun Long.toContactId(): ContactId? =
     }
 
 @JvmInline
-value class ContactId(override val value: Long): DashboardItemId {
+value class ContactId(
+    override val value: Long
+    ): DashboardItemId {
     companion object {
         const val NULL_CONTACT_ID = Long.MAX_VALUE
     }