+++ sphinx/application/common/wrappers/wrapper-common/src/main/java/chat/sphinx/wrapper_common/subscription/Cron.kt
@@ -1,0 +1,10 @@
+package chat.sphinx.wrapper_common.subscription
+
+@JvmInline
+value class Cron(val value: String) {
+    init {
+        require(value.isNotEmpty()) {
+            "Subscription Cron cannot be empty"
+        }
+    }
+}