+++ sphinx/application/data/features/feature-repository/src/main/java/chat/sphinx/feature_repository/SphinxRepository.kt
@@ -20,6 +20,10 @@ import chat.sphinx.concept_network_query_message.NetworkQueryMessage
 import chat.sphinx.concept_network_query_message.model.MessageDto
 import chat.sphinx.concept_network_query_message.model.PostMessageDto
 import chat.sphinx.concept_network_query_message.model.PostPaymentDto
+import chat.sphinx.concept_network_query_subscription.NetworkQuerySubscription
+import chat.sphinx.concept_network_query_subscription.model.PostSubscriptionDto
+import chat.sphinx.concept_network_query_subscription.model.PutSubscriptionDto
+import chat.sphinx.concept_network_query_subscription.model.SubscriptionDto
 import chat.sphinx.concept_network_query_verify_external.NetworkQueryAuthorizeExternal
 import chat.sphinx.concept_repository_chat.ChatRepository
 import chat.sphinx.concept_repository_chat.model.CreateTribe
@@ -32,6 +36,7 @@ import chat.sphinx.concept_repository_message.MessageRepository
 import chat.sphinx.concept_repository_message.model.AttachmentInfo
 import chat.sphinx.concept_repository_message.model.SendMessage
 import chat.sphinx.concept_repository_message.model.SendPayment
+import chat.sphinx.concept_repository_subscription.SubscriptionRepository
 import chat.sphinx.concept_socket_io.SocketIOManager
 import chat.sphinx.concept_socket_io.SphinxSocketIOMessage
 import chat.sphinx.concept_socket_io.SphinxSocketIOMessageListener
@@ -41,6 +46,7 @@ import chat.sphinx.feature_repository.mappers.contact.ContactDboPresenterMapper
 import chat.sphinx.feature_repository.mappers.invite.InviteDboPresenterMapper
 import chat.sphinx.feature_repository.mappers.mapListFrom
 import chat.sphinx.feature_repository.mappers.message.MessageDboPresenterMapper
+import chat.sphinx.feature_repository.mappers.subscription.SubscriptionDboPresenterMapper
 import chat.sphinx.feature_repository.model.MessageDboWrapper
 import chat.sphinx.feature_repository.model.MessageMediaDboWrapper
 import chat.sphinx.feature_repository.util.*
@@ -59,6 +65,8 @@ import chat.sphinx.wrapper_common.dashboard.toChatId
 import chat.sphinx.wrapper_common.invite.InviteStatus
 import chat.sphinx.wrapper_common.lightning.*
 import chat.sphinx.wrapper_common.message.*
+import chat.sphinx.wrapper_common.subscription.EndNumber
+import chat.sphinx.wrapper_common.subscription.SubscriptionId
 import chat.sphinx.wrapper_contact.*
 import chat.sphinx.wrapper_invite.Invite
 import chat.sphinx.wrapper_io_utils.InputStreamProvider
@@ -68,12 +76,14 @@ import chat.sphinx.wrapper_meme_server.PublicAttachmentInfo
 import chat.sphinx.wrapper_message.*
 import chat.sphinx.wrapper_message_media.*
 import chat.sphinx.wrapper_message_media.token.MediaHost
+import chat.sphinx.wrapper_message_media.token.MediaMUID
 import chat.sphinx.wrapper_podcast.Podcast
 import chat.sphinx.wrapper_podcast.PodcastDestination
 import chat.sphinx.wrapper_relay.AuthorizationToken
 import chat.sphinx.wrapper_relay.RelayUrl
 import chat.sphinx.wrapper_rsa.RsaPrivateKey
 import chat.sphinx.wrapper_rsa.RsaPublicKey
+import chat.sphinx.wrapper_subscription.Subscription
 import com.squareup.moshi.Moshi
 import com.squareup.sqldelight.runtime.coroutines.asFlow
 import com.squareup.sqldelight.runtime.coroutines.mapToList
@@ -109,6 +119,7 @@ abstract class SphinxRepository(
     private val networkQueryMessage: NetworkQueryMessage,
     private val networkQueryInvite: NetworkQueryInvite,
     private val networkQueryAuthorizeExternal: NetworkQueryAuthorizeExternal,
+    private val networkQuerySubscription: NetworkQuerySubscription,
     private val rsa: RSA,
     private val socketIOManager: SocketIOManager,
     protected val LOG: SphinxLogger,
@@ -116,6 +127,7 @@ abstract class SphinxRepository(
     ContactRepository,
     LightningRepository,
     MessageRepository,
+    SubscriptionRepository,
     RepositoryDashboard,
     RepositoryMedia,
     CoroutineDispatchers by dispatchers,
@@ -1447,6 +1459,7 @@ abstract class SphinxRepository(
         queries: SphinxDatabaseQueries,
         messageDbo: MessageDbo,
         reactions: List<Message>? = null,
+        purchaseItems: List<Message>? = null,
         replyMessage: ReplyUUID? = null,
     ): Message {
 
@@ -1574,6 +1587,7 @@ abstract class SphinxRepository(
         }
 
         message._reactions = reactions
+        message._purchaseItems = purchaseItems
 
         replyMessage?.value?.toMessageUUID()?.let { uuid ->
             queries.messageGetToShowByUUID(uuid).executeAsOneOrNull()?.let { replyDbo ->
@@ -1596,13 +1610,21 @@ abstract class SphinxRepository(
                         val reactionsMap: MutableMap<MessageUUID, ArrayList<Message>> =
                             LinkedHashMap(listMessageDbo.size)
 
+                        val purchaseItemsMap: MutableMap<MessageMUID, ArrayList<Message>> =
+                            LinkedHashMap(listMessageDbo.size)
+
                         for (dbo in listMessageDbo) {
                             dbo.uuid?.let { uuid ->
                                 reactionsMap[uuid] = ArrayList(0)
                             }
+                            dbo.muid?.let { muid ->
+                                purchaseItemsMap[muid] = ArrayList(0)
+                            }
                         }
 
                         val replyUUIDs = reactionsMap.keys.map { ReplyUUID(it.value) }
+
+                        val purchaseItemsMUIDs = purchaseItemsMap.keys.map { MessageMUID(it.value) }
 
                         replyUUIDs.chunked(500).forEach { chunkedIds ->
                             queries.messageGetAllReactionsByUUID(
@@ -1620,11 +1642,28 @@ abstract class SphinxRepository(
                                 }
                         }
 
+                        purchaseItemsMUIDs.chunked(500).forEach { chunkedMUIDs ->
+                            queries.messageGetAllPurchaseItemsByMUID(
+                                chatId,
+                                chunkedMUIDs,
+                            ).executeAsList()
+                                .let { response ->
+                                    response.forEach { dbo ->
+                                        dbo.muid?.let { muid ->
+                                            purchaseItemsMap[muid]?.add(
+                                                mapMessageDboAndDecryptContentIfNeeded(queries, dbo)
+                                            )
+                                        }
+                                    }
+                                }
+                        }
+
                         listMessageDbo.reversed().map { dbo ->
                             mapMessageDboAndDecryptContentIfNeeded(
                                 queries,
                                 dbo,
                                 dbo.uuid?.let { reactionsMap[it] },
+                                dbo.muid?.let { purchaseItemsMap[it] },
                                 dbo.reply_uuid,
                             )
                         }
@@ -1899,6 +1938,7 @@ abstract class SphinxRepository(
                                 null,
                                 message?.second,
                                 message?.first,
+                                null
                             )
 
                             if (media != null) {
@@ -1987,6 +2027,7 @@ abstract class SphinxRepository(
                     mediaKeyMap,
                     postMemeServerDto?.mime,
                     postMemeServerDto?.muid,
+                    sendMessage.messagePrice?.value,
                     sendMessage.isBoost
                 )
             } catch (e: IllegalArgumentException) {
@@ -2232,6 +2273,7 @@ abstract class SphinxRepository(
                     null,
                     message.messageMedia?.mediaType?.value,
                     message.messageMedia?.muid?.value,
+                    null,
                     false
                 )
             } catch (e: IllegalArgumentException) {
@@ -2636,6 +2678,51 @@ abstract class SphinxRepository(
         }.join()
 
         return response ?: Response.Error(ResponseError(""))
+    }
+
+    override suspend fun payAttachment(message: Message) : Response<Any, ResponseError> {
+        var response: Response<Any, ResponseError> = Response.Error(ResponseError("Failed to pay for attachment"))
+
+        applicationScope.launch(mainImmediate) {
+            val queries = coreDB.getSphinxDatabaseQueries()
+
+            message.messageMedia?.mediaToken?.let { mediaToken ->
+                mediaToken.getPriceFromMediaToken()?.let { price ->
+
+                    networkQueryMessage.payAttachment(
+                        message.chatId,
+                        message.sender,
+                        price,
+                        mediaToken
+                    ).collect { loadResponse ->
+                        @Exhaustive
+                        when (loadResponse) {
+                            is LoadResponse.Loading -> {}
+
+                            is Response.Error -> {
+                                response = Response.Error(
+                                    ResponseError(loadResponse.message, loadResponse.exception)
+                                )
+                            }
+                            is Response.Success -> {
+                                response = loadResponse
+
+                                messageLock.withLock {
+                                    withContext(io) {
+                                        queries.transaction {
+                                            upsertMessage(loadResponse.value, queries)
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                }
+            }
+        }.join()
+
+        return response
     }
 
     override suspend fun toggleChatMuted(chat: Chat): Response<Boolean, ResponseError> {
@@ -3549,5 +3636,235 @@ abstract class SphinxRepository(
         }.join()
 
         return response
+    }
+
+    /***
+     * Subscriptions
+     */
+
+    private val subscriptionLock = Mutex()
+    private val subscriptionDboPresenterMapper: SubscriptionDboPresenterMapper by lazy {
+        SubscriptionDboPresenterMapper(dispatchers)
+    }
+
+    override fun getActiveSubscriptionByContactId(contactId: ContactId): Flow<Subscription?> = flow {
+        emitAll(
+            coreDB.getSphinxDatabaseQueries().subscriptionGetLastActiveByContactId(contactId)
+                .asFlow()
+                .mapToOneOrNull(io)
+                .map { it?.let { subscriptionDboPresenterMapper.mapFrom(it) } }
+                .distinctUntilChanged()
+        )
+    }
+
+    override suspend fun createSubscription(
+        amount: Sat,
+        interval: String,
+        contactId: ContactId,
+        chatId: ChatId?,
+        endDate: String?,
+        endNumber: EndNumber?
+    ): Response<Any, ResponseError> {
+        var response: Response<SubscriptionDto, ResponseError>? = null
+
+        applicationScope.launch(mainImmediate) {
+            networkQuerySubscription.postSubscription(
+                PostSubscriptionDto(
+                    amount = amount.value,
+                    contact_id = contactId.value,
+                    chat_id = chatId?.value,
+                    interval = interval,
+                    end_number = endNumber?.value,
+                    end_date = endDate
+                )
+            ).collect { loadResponse ->
+                @Exhaustive
+                when (loadResponse) {
+                    is LoadResponse.Loading -> {}
+                    is Response.Error -> {
+                        response = loadResponse
+                    }
+                    is Response.Success -> {
+                        response = loadResponse
+                        val queries = coreDB.getSphinxDatabaseQueries()
+
+                        subscriptionLock.withLock {
+                            withContext(io) {
+                                queries.transaction {
+                                    upsertSubscription(
+                                        loadResponse.value,
+                                        queries
+                                    )
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }.join()
+
+        return response ?: Response.Error(ResponseError(("Failed to create subscription")))
+    }
+
+    override suspend fun updateSubscription(
+        id: SubscriptionId,
+        amount: Sat,
+        interval: String,
+        contactId: ContactId,
+        chatId: ChatId?,
+        endDate: String?,
+        endNumber: EndNumber?
+    ): Response<Any, ResponseError> {
+        var response: Response<SubscriptionDto, ResponseError>? = null
+
+        applicationScope.launch(mainImmediate) {
+
+            networkQuerySubscription.putSubscription(
+                id,
+                PutSubscriptionDto(
+                    amount = amount.value,
+                    contact_id = contactId.value,
+                    chat_id = chatId?.value,
+                    interval = interval,
+                    end_number = endNumber?.value,
+                    end_date = endDate
+                )
+            ).collect { loadResponse ->
+                @Exhaustive
+                when (loadResponse) {
+                    is LoadResponse.Loading -> {}
+                    is Response.Error -> {
+                        response = loadResponse
+                    }
+                    is Response.Success -> {
+                        response = loadResponse
+                        val queries = coreDB.getSphinxDatabaseQueries()
+
+                        subscriptionLock.withLock {
+                            withContext(io) {
+                                queries.transaction {
+                                    upsertSubscription(
+                                        loadResponse.value,
+                                        queries
+                                    )
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }.join()
+
+        return response ?: Response.Error(ResponseError(("Failed to update subscription")))
+    }
+
+    override suspend fun restartSubscription(
+        subscriptionId: SubscriptionId
+    ): Response<Any, ResponseError> {
+        var response: Response<SubscriptionDto, ResponseError>? = null
+
+        applicationScope.launch(mainImmediate) {
+
+            networkQuerySubscription.putRestartSubscription(
+                subscriptionId
+            ).collect { loadResponse ->
+                @Exhaustive
+                when (loadResponse) {
+                    is LoadResponse.Loading -> {}
+                    is Response.Error -> {
+                        response = loadResponse
+                    }
+                    is Response.Success -> {
+                        response = loadResponse
+                        val queries = coreDB.getSphinxDatabaseQueries()
+
+                        subscriptionLock.withLock {
+                            withContext(io) {
+                                queries.transaction {
+                                    upsertSubscription(
+                                        loadResponse.value,
+                                        queries
+                                    )
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }.join()
+
+        return response ?: Response.Error(ResponseError(("Failed to restart subscription")))
+    }
+
+    override suspend fun pauseSubscription(
+        subscriptionId: SubscriptionId
+    ): Response<Any, ResponseError> {
+        var response: Response<SubscriptionDto, ResponseError>? = null
+
+        applicationScope.launch(mainImmediate) {
+
+            networkQuerySubscription.putPauseSubscription(
+                subscriptionId
+            ).collect { loadResponse ->
+                @Exhaustive
+                when (loadResponse) {
+                    is LoadResponse.Loading -> {}
+                    is Response.Error -> {
+                        response = loadResponse
+                    }
+                    is Response.Success -> {
+                        response = loadResponse
+                        val queries = coreDB.getSphinxDatabaseQueries()
+
+                        subscriptionLock.withLock {
+                            withContext(io) {
+                                queries.transaction {
+                                    upsertSubscription(
+                                        loadResponse.value,
+                                        queries
+                                    )
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }.join()
+
+        return response ?: Response.Error(ResponseError(("Failed to pause subscription")))
+    }
+
+    override suspend fun deleteSubscription(
+        subscriptionId: SubscriptionId
+    ): Response<Any, ResponseError> {
+        var response: Response<Any, ResponseError>? = null
+
+        applicationScope.launch(mainImmediate) {
+            networkQuerySubscription.deleteSubscription(
+                subscriptionId
+            ).collect { loadResponse ->
+                @Exhaustive
+                when (loadResponse) {
+                    is LoadResponse.Loading -> {}
+                    is Response.Error -> {
+                        response = loadResponse
+                    }
+                    is Response.Success -> {
+                        response = loadResponse
+                        val queries = coreDB.getSphinxDatabaseQueries()
+
+                        subscriptionLock.withLock {
+                            withContext(io) {
+                                queries.transaction {
+                                    deleteSubscriptionById(subscriptionId, queries)
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }.join()
+
+        return response ?: Response.Error(ResponseError(("Failed to delete subscription")))
     }
 }