+++ sphinx/application/data/features/feature-repository/src/main/java/chat/sphinx/feature_repository/util/Extensions.kt
@@ -6,11 +6,12 @@ import chat.sphinx.concept_network_query_contact.model.ContactDto
 import chat.sphinx.concept_network_query_invite.model.InviteDto
 import chat.sphinx.concept_network_query_lightning.model.balance.BalanceDto
 import chat.sphinx.concept_network_query_message.model.MessageDto
+import chat.sphinx.concept_network_query_subscription.model.SubscriptionDto
 import chat.sphinx.conceptcoredb.SphinxDatabaseQueries
 import chat.sphinx.wrapper_chat.*
 import chat.sphinx.wrapper_common.*
-import chat.sphinx.wrapper_common.dashboard.ChatId
 import chat.sphinx.wrapper_common.chat.ChatUUID
+import chat.sphinx.wrapper_common.dashboard.ChatId
 import chat.sphinx.wrapper_common.dashboard.ContactId
 import chat.sphinx.wrapper_common.dashboard.InviteId
 import chat.sphinx.wrapper_common.invite.InviteStatus
@@ -20,14 +21,15 @@ import chat.sphinx.wrapper_common.invite.toInviteStatus
 import chat.sphinx.wrapper_common.lightning.*
 import chat.sphinx.wrapper_common.message.MessageId
 import chat.sphinx.wrapper_common.message.toMessageUUID
+import chat.sphinx.wrapper_common.subscription.Cron
+import chat.sphinx.wrapper_common.subscription.EndNumber
+import chat.sphinx.wrapper_common.subscription.SubscriptionCount
+import chat.sphinx.wrapper_common.subscription.SubscriptionId
 import chat.sphinx.wrapper_contact.*
 import chat.sphinx.wrapper_invite.InviteString
 import chat.sphinx.wrapper_lightning.NodeBalance
 import chat.sphinx.wrapper_message.*
-import chat.sphinx.wrapper_message_media.MediaToken
-import chat.sphinx.wrapper_message_media.toMediaKey
-import chat.sphinx.wrapper_message_media.toMediaKeyDecrypted
-import chat.sphinx.wrapper_message_media.toMediaType
+import chat.sphinx.wrapper_message_media.*
 import chat.sphinx.wrapper_rsa.RsaPublicKey
 import com.squareup.moshi.Moshi
 import com.squareup.sqldelight.TransactionCallbacks
@@ -325,21 +327,19 @@ fun TransactionCallbacks.upsertMessage(dto: MessageDto, queries: SphinxDatabaseQ
     } ?: ChatId(ChatId.NULL_CHAT_ID.toLong())
 
     dto.media_token?.let { mediaToken ->
-        dto.media_type?.let { mediaType ->
 
-            if (mediaToken.isEmpty() || mediaType.isEmpty()) return
+        if (mediaToken.isEmpty()) return
 
-            queries.messageMediaUpsert(
-                (dto.media_key ?: "").toMediaKey(),
-                mediaType.toMediaType(),
-                MediaToken(mediaToken),
-                MessageId(dto.id),
-                chatId,
-                dto.mediaKeyDecrypted?.toMediaKeyDecrypted(),
-                dto.mediaLocalFile,
-            )
+        queries.messageMediaUpsert(
+            (dto.media_key ?: "").toMediaKey(),
+            (dto.media_type ?: "").toMediaType(),
+            MediaToken(mediaToken),
+            MessageId(dto.id),
+            chatId,
+            dto.mediaKeyDecrypted?.toMediaKeyDecrypted(),
+            dto.mediaLocalFile,
+        )
 
-        }
     }
 
     queries.messageUpsert(
@@ -362,24 +362,8 @@ fun TransactionCallbacks.upsertMessage(dto: MessageDto, queries: SphinxDatabaseQ
         dto.expiration_date?.toDateTime(),
         dto.message_content?.toMessageContent(),
         dto.messageContentDecrypted?.toMessageContentDecrypted(),
+        dto.media_token?.toMediaToken()?.getMUIDFromMediaToken()?.value?.toMessageMUID()
     )
-
-    dto.media_token?.let { mediaToken ->
-        dto.media_type?.let { mediaType ->
-
-            if (mediaToken.isEmpty() || mediaType.isEmpty()) return
-
-            queries.messageMediaUpsert(
-                (dto.media_key ?: "").toMediaKey(),
-                mediaType.toMediaType(),
-                MediaToken(mediaToken),
-                MessageId(dto.id),
-                chatId,
-                dto.mediaKeyDecrypted?.toMediaKeyDecrypted(),
-                dto.mediaLocalFile
-            )
-        }
-    }
 }
 
 @Suppress("NOTHING_TO_INLINE")
@@ -411,6 +395,7 @@ inline fun TransactionCallbacks.deleteContactById(
     queries.contactDeleteById(contactId)
     queries.inviteDeleteByContactId(contactId)
     queries.dashboardDeleteById(contactId)
+    queries.subscriptionDeleteByContactId(contactId)
 }
 
 @Suppress("NOTHING_TO_INLINE")
@@ -420,4 +405,30 @@ inline fun TransactionCallbacks.deleteMessageById(
 ) {
     queries.messageDeleteById(messageId)
     queries.messageMediaDeleteById(messageId)
+}
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun TransactionCallbacks.upsertSubscription(subscriptionDto: SubscriptionDto, queries: SphinxDatabaseQueries) {
+    queries.subscriptionUpsert(
+        id = SubscriptionId(subscriptionDto.id),
+        amount = Sat(subscriptionDto.amount),
+        contact_id = ContactId(subscriptionDto.contact_id),
+        chat_id = ChatId(subscriptionDto.chat_id),
+        count = SubscriptionCount(subscriptionDto.count.toLong()),
+        cron = Cron(subscriptionDto.cron),
+        end_date = subscriptionDto.end_date?.toDateTime(),
+        end_number = subscriptionDto.end_number?.let { EndNumber(it.toLong()) },
+        created_at = subscriptionDto.created_at.toDateTime(),
+        updated_at = subscriptionDto.updated_at.toDateTime(),
+        ended = subscriptionDto.endedActual,
+        paused = subscriptionDto.pausedActual,
+    )
+}
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun TransactionCallbacks.deleteSubscriptionById(
+    subscriptionId: SubscriptionId,
+    queries: SphinxDatabaseQueries
+) {
+    queries.subscriptionDeleteById(subscriptionId)
 }