+++ sphinx/application/network/concepts/queries/concept-network-query-message/src/main/java/chat/sphinx/concept_network_query_message/NetworkQueryMessage.kt
@@ -11,6 +11,7 @@ import chat.sphinx.wrapper_common.message.MessagePagination
 import chat.sphinx.wrapper_common.message.MessageUUID
 import chat.sphinx.wrapper_message.MessageStatus
 import chat.sphinx.wrapper_message.MessageType
+import chat.sphinx.wrapper_message_media.MediaToken
 import chat.sphinx.wrapper_relay.AuthorizationToken
 import chat.sphinx.wrapper_relay.RelayUrl
 import kotlinx.coroutines.flow.Flow
@@ -52,6 +53,14 @@ abstract class NetworkQueryMessage {
         relayData: Pair<AuthorizationToken, RelayUrl>? = null,
     ): Flow<LoadResponse<KeySendPaymentDto, ResponseError>>
 
+    abstract fun payAttachment(
+        chatId: ChatId,
+        contactId: ContactId?,
+        amount: Sat,
+        mediaToken: MediaToken,
+        relayData: Pair<AuthorizationToken, RelayUrl>? = null
+    ): Flow<LoadResponse<MessageDto, ResponseError>>
+
     abstract fun readMessages(
         chatId: ChatId,
         relayData: Pair<AuthorizationToken, RelayUrl>? = null,
@@ -78,5 +87,4 @@ abstract class NetworkQueryMessage {
         type: MessageType,
         relayData: Pair<AuthorizationToken, RelayUrl>? = null
     ): Flow<LoadResponse<PutMemberResponseDto, ResponseError>>
-
 }