+++ sphinx/application/network/features/queries/feature-network-query-subscription/src/main/java/chat/sphinx/feature_network_query_subscription/model/SubscriptionRelayResponse.kt
@@ -5,7 +5,7 @@ import chat.sphinx.concept_network_relay_call.RelayResponse
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class GetSubscriptionRelayResponse(
+data class SubscriptionRelayResponse(
     override val success: Boolean,
     override val response: SubscriptionDto?,
     override val error: String?