+++ sphinx/screens-detail/contact/edit-contact/src/main/java/chat/sphinx/edit_contact/ui/EditContactFragment.kt
@@ -1,14 +1,18 @@
 package chat.sphinx.edit_contact.ui
 
+import android.os.Bundle
+import android.view.View
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.lifecycleScope
 import by.kirich1409.viewbindingdelegate.viewBinding
 import chat.sphinx.concept_user_colors_helper.UserColorsHelper
 import chat.sphinx.contact.databinding.LayoutContactBinding
+import chat.sphinx.contact.databinding.LayoutContactDetailScreenHeaderBinding
 import chat.sphinx.contact.ui.ContactFragment
-import chat.sphinx.detail_resources.databinding.LayoutDetailScreenHeaderBinding
 import chat.sphinx.edit_contact.R
 import chat.sphinx.edit_contact.databinding.FragmentEditContactBinding
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -27,8 +31,8 @@ internal class EditContactFragment : ContactFragment<
     override val viewModel: EditContactViewModel by viewModels()
     override val binding: FragmentEditContactBinding by viewBinding(FragmentEditContactBinding::bind)
 
-    override val headerBinding: LayoutDetailScreenHeaderBinding by viewBinding(
-        LayoutDetailScreenHeaderBinding::bind, R.id.include_edit_contact_header
+    override val headerBinding: LayoutContactDetailScreenHeaderBinding by viewBinding(
+        LayoutContactDetailScreenHeaderBinding::bind, R.id.include_edit_contact_header
     )
     override val contactBinding: LayoutContactBinding by viewBinding(
         LayoutContactBinding::bind, R.id.include_edit_contact_layout
@@ -37,4 +41,16 @@ internal class EditContactFragment : ContactFragment<
     override fun getHeaderText(): String = getString(R.string.edit_contact_header_name)
 
     override fun getSaveButtonText(): String = getString(R.string.save_contact_button)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        headerBinding.apply {
+            textViewDetailScreenSubscribe.setOnClickListener {
+                lifecycleScope.launch(viewModel.mainImmediate) {
+                    viewModel.toSubscriptionDetailScreen()
+                }
+            }
+        }
+    }
 }